Refactored Summary class into it's own module
